[skywallet] refs #112 - Rename device-wallet => skywallet
